fixed the styling bugs
my error was that I was overriding my override by trying to use the _custom.scss in mulitply places.  Only one file can be used to override the bootstrap.scss and those overrides have to all be dones in _custom.scss. My error was to try to make modular .scss files for customizing the _custom.scss file.