carousel and control via thumbails of carousel as well as active, and hover states working
made a custom JS script to handle the behaviour of the active states on the thumbnail gallery. I could not find in the docs a way to do this with Bootstrap utilies or components, nor did I have much luck google searching for solutions. Next I need to hook up the carousel state to the thumb galleries, so when you navigate via the carousel, the active states on the thumbnails is dynamic