starting scaffing out product info card.
Also starting refactory how I handle custom css and bootstraps frame work. Goal is to trim back on as much custom code as possible and make more use of Bootstrap utilies and helpers now that I have a firmer understanding of how to work witn the framework. And to provide a code base and system that others could use with a relavity fast aqquanitance and undertstanding time.